Move JTC action server and trajectory sampling into separate class
